requirements.txt
biopython==1.83
numpy==1.26.3
